requirements.txt
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.1
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2